backend/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.0.5
Flask-Migrate>=4.0.5
Flask-JWT-Extended>=4.6.0
Flask-CORS>=4.0.0
python-dotenv>=1.0.0
bcrypt>=4.1.2
marshmallow>=3.20.2
marshmallow-sqlalchemy>=1.0.0
Flask-Limiter>=3.5.0
flasgger>=0.9.7.1
Pillow>=10.0.0
bleach>=6.1.0
gunicorn>=21.2.0
Werkzeug>=3.0.1
email-validator>=2.1.0
pytest>=7.4.0
pytest-flask>=1.3.0
